refactoring cache path methods into RequestUtils for use in other classes
